# apps/metadata-server/build/deploy/swagger.yaml
...
swagger: '2.0'
info:
  version: 2.0.0
  title: GAIA Light Metadata Server $CONTAINER_CLUSTER $GCLOUD_PROJECT
  description: GAIA Light server for providing metadata. Data is persisted in JSON files.
host: metadata-server-$CONTAINER_CLUSTER.endpoints.$GCLOUD_PROJECT.cloud.goog
basePath: /api
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /metadata/map-config:
    get:
      operationId: getMapConfiguration
      tags:
        - map
      summary: 'map configuration '
      description: 'map configuration endpoint '
      parameters:
        - in: header
          name: request
          type: string
          description: 'incoming request object'
          required: true
      responses:
        '200':
          description: returns map configuration for the given request host
        '500':
          description: returns when the request host is not found in the configuration
...
  /metadata/setting:
    post:
      operationId: postLayerSetting
      parameters:
        - name: Authorization
          in: header
          required: true
          default: Bearer {token}
          type: string
        - name: layerSetting
          in: body
          schema:
            type: object
      tags:
        - Map
      summary: Update existing layer if existing or create new layer.
      description: Update existing layer if existing or create new layer.
      responses:
        '200':
          description: ''
          headers: {}
...